# Improve State and rewrite step functions