--- a/server/interpreter.js
+++ b/server/interpreter.js
@@ -4736,8 +4736,8 @@ stepFuncs_['EmptyStatement'] = function (stack, state, node) {
  * @return {!Interpreter.State|undefined}
  */
 stepFuncs_['EvalProgram_'] = function (stack, state, node) {
-  var n = state.n_ || 0;
-  var expression = node['body'][n];
+  var n = state.n_;
+  var /** ?Interpreter.Node */ expression = node['body'][n];
   if (expression) {
     state.n_ = n + 1;
     return new Interpreter.State(expression, state.scope);