The impetus was to improve improve type safety in step functions by making sure all State instance properties were declared.  All the step functions were rewritten to avoid mutating the shape of State instances, and improve typing and readability.

A performance improvement was expected, but that it turns out to be about 20% is a bonus.

Before:
```
WARMUP	fibonacci10k...
WARMUP	fibonacci10k: 3553 ms
RUN 1	fibonacci10k...
RUN 1	fibonacci10k: 3495 ms
RUN 2	fibonacci10k...
RUN 2	fibonacci10k: 3452 ms
RUN 3	fibonacci10k...
RUN 3	fibonacci10k: 3428 ms
WARMUP	ressurrectedFibonacci10k...
WARMUP	ressurrectedFibonacci10k: 3536 ms
RUN 1	ressurrectedFibonacci10k...
RUN 1	ressurrectedFibonacci10k: 3634 ms
RUN 2	ressurrectedFibonacci10k...
RUN 2	ressurrectedFibonacci10k: 3468 ms
RUN 3	ressurrectedFibonacci10k...
RUN 3	ressurrectedFibonacci10k: 3510 ms
```
After:
```
WARMUP	fibonacci10k...
WARMUP	fibonacci10k: 2787 ms
RUN 1	fibonacci10k...
RUN 1	fibonacci10k: 2755 ms
RUN 2	fibonacci10k...
RUN 2	fibonacci10k: 2856 ms
RUN 3	fibonacci10k...
RUN 3	fibonacci10k: 2740 ms
WARMUP	ressurrectedFibonacci10k...
WARMUP	ressurrectedFibonacci10k: 2824 ms
RUN 1	ressurrectedFibonacci10k...
RUN 1	ressurrectedFibonacci10k: 2816 ms
RUN 2	ressurrectedFibonacci10k...
RUN 2	ressurrectedFibonacci10k: 2859 ms
RUN 3	ressurrectedFibonacci10k...
RUN 3	ressurrectedFibonacci10k: 2839 ms
```